refactor: move event parsing logic to a new module
