+++ chain-signatures/chain-ethereum/src/event_parsing.rs
@@ -1,0 +1,264 @@
+//! Decoding of `ChainSignatures` contract events and helper predicates used by
+//! the Ethereum indexer.
+
+use crate::abi::{ChainSignatures, SignatureRequestedEncoding};
+use alloy::primitives::hex::{self, ToHexExt};
+use alloy::primitives::{Address, Bytes, U256};
+use alloy::rpc::types::Log;
+use alloy::sol_types::SolEvent;
+use k256::elliptic_curve::sec1::FromEncodedPoint;
+use k256::{AffinePoint as K256AffinePoint, EncodedPoint, FieldBytes, Scalar};
+use mpc_crypto::{kdf::derive_epsilon_eth, ScalarExt as _};
+use mpc_primitives::{
+    Chain, ChainEvent, IndexedSignRequest, SignArgs, SignId, Signature as MpcSignature,
+    SignatureRespondedEvent, LATEST_MPC_KEY_VERSION, MAX_SECP256K1_SCALAR,
+};
+use tokio::sync::mpsc;
+
+/// Whether a transaction's calldata represents a contract call.
+pub fn is_contract_call(input: &Bytes) -> bool {
+    input.len() > 2 && input != &Bytes::from("0x")
+}
+
+pub fn parse_filtered_logs(logs: Vec<Log>) -> Vec<IndexedSignRequest> {
+    let mut indexed_requests = Vec::new();
+    for log in logs {
+        tracing::debug!("Parsing Ethereum log: {:?}", log);
+        match sign_request_from_filtered_log(log.clone()) {
+            Some(request) => indexed_requests.push(request),
+            None => tracing::warn!("Failed to parse Ethereum log: {:?}", log),
+        }
+    }
+    if indexed_requests.is_empty() {
+        tracing::warn!("No valid Ethereum sign requests found in logs");
+    }
+    indexed_requests
+}
+
+pub async fn emit_respond_events(logs: &[Log], events_tx: mpsc::Sender<ChainEvent>) {
+    for log in logs {
+        let Some(sign_id) = sign_id_from_signature_responded_log(log) else {
+            continue;
+        };
+
+        let data = &log.data().data;
+        if data.len() < 160 {
+            tracing::warn!(
+                ?sign_id,
+                data_len = data.len(),
+                "signature event data too short to parse full signature: skipping..."
+            );
+            continue;
+        }
+
+        // signature struct encoding layout:
+        // bigR.x at 32..64, bigR.y at 64..96, s at 96..128, recoveryId at 159
+        let big_r_x = &data[32..64];
+        let big_r_y = &data[64..96];
+        let s_bytes: [u8; 32] = data[96..128].try_into().unwrap();
+        let recovery_id = data[159];
+
+        let x_field = FieldBytes::from_slice(big_r_x);
+        let y_field = FieldBytes::from_slice(big_r_y);
+        let encoded_r = EncodedPoint::from_affine_coordinates(x_field, y_field, false);
+        let Some(big_r) = K256AffinePoint::from_encoded_point(&encoded_r).into_option() else {
+            tracing::warn!(?sign_id, "ethereum respond event, invalid big_r point");
+            continue;
+        };
+
+        let Some(s) = Scalar::from_bytes(s_bytes) else {
+            tracing::warn!(?sign_id, "ethereum respond event, invalid s scalar");
+            continue;
+        };
+
+        let signature = MpcSignature::new(big_r, s, recovery_id);
+
+        let respond_event = SignatureRespondedEvent {
+            request_id: sign_id.request_id,
+            signature,
+            chain: Chain::Ethereum,
+        };
+        tracing::info!(?sign_id, "emitting SignatureResponded event");
+        if let Err(err) = events_tx.send(ChainEvent::Respond(respond_event)).await {
+            tracing::error!(?err, "failed to emit Respond event");
+        }
+    }
+}
+
+fn sign_id_from_signature_responded_log(log: &Log) -> Option<SignId> {
+    if log
+        .topic0()
+        .is_none_or(|topic| *topic != ChainSignatures::SignatureResponded::SIGNATURE_HASH)
+    {
+        return None;
+    }
+
+    let request_topic = log.topics().get(1)?;
+    let request_id: [u8; 32] = (*request_topic).into();
+    Some(SignId { request_id })
+}
+
+fn sign_request_from_filtered_log(log: Log) -> Option<IndexedSignRequest> {
+    let event = parse_event(&log);
+    tracing::debug!("found eth event: {:?}", event);
+    if event.deposit == U256::ZERO {
+        tracing::warn!("deposit is 0, skipping sign request");
+        return None;
+    }
+
+    if event.key_version > LATEST_MPC_KEY_VERSION {
+        tracing::warn!("unsupported key version: {}", event.key_version);
+        return None;
+    }
+
+    // Create sign request from event
+    let Some(payload) = Scalar::from_bytes(event.payload_hash) else {
+        tracing::warn!(
+            "eth `sign` did not produce payload hash correctly: {:?}",
+            event.payload_hash,
+        );
+        return None;
+    };
+
+    if payload > *MAX_SECP256K1_SCALAR {
+        tracing::warn!("payload exceeds secp256k1 curve order: {payload:?}");
+        return None;
+    }
+
+    let epsilon = derive_epsilon_eth(
+        event.key_version,
+        format!("0x{}", event.requester.encode_hex()).as_str(),
+        &event.path,
+    );
+
+    // Use transaction hash as entropy
+    let entropy = log.transaction_hash.unwrap_or_default();
+
+    let sign_id = SignId::new(event.generate_request_id());
+    tracing::info!(?sign_id, "eth signature requested");
+
+    Some(IndexedSignRequest::sign(
+        sign_id,
+        SignArgs {
+            entropy: entropy.into(),
+            epsilon,
+            payload,
+            path: event.path,
+            key_version: event.key_version,
+        },
+        Chain::Ethereum,
+        crate::util::current_unix_timestamp(),
+    ))
+}
+
+// Helper function to parse event logs
+fn parse_event(log: &Log) -> SignatureRequestedEvent {
+    // Parse data fields
+    let data = log.data().data.clone();
+
+    // Parse requester address (20 bytes)
+    let requester = Address::from_slice(&data[12..32]);
+
+    // Parse payload hash (32 bytes)
+    let mut payload_hash = [0u8; 32];
+    payload_hash.copy_from_slice(&data[32..64]);
+
+    let key_version: u32 = U256::from_be_slice(&data[64..96]).to::<u32>();
+
+    let deposit = U256::from_be_slice(&data[96..128]);
+
+    let chain_id = U256::from_be_slice(&data[128..160]);
+
+    let path = parse_string_args(&data, 160);
+
+    let algo = parse_string_args(&data, 192);
+
+    let dest = parse_string_args(&data, 224);
+
+    let params = parse_string_args(&data, 256);
+
+    tracing::info!(
+        "Parsed event: requester={}, payload_hash={}, path={}, deposit={}, chain_id={}, algo={}, dest={}, params={}",
+        requester,
+        hex::encode(payload_hash),
+        path,
+        deposit,
+        chain_id,
+        algo,
+        dest,
+        params
+    );
+
+    SignatureRequestedEvent {
+        requester,
+        payload_hash,
+        path,
+        key_version,
+        deposit,
+        chain_id,
+        algo,
+        dest,
+        params,
+    }
+}
+
+fn parse_string_args(data: &Bytes, offset_start: usize) -> String {
+    let offset: usize = U256::from_be_slice(&data[offset_start..offset_start + 32]).to::<usize>();
+    let length: usize = U256::from_be_slice(&data[offset..offset + 32]).to::<usize>();
+    if length == 0 {
+        return String::new();
+    }
+    let bytes = &data[offset + 32..offset + 32 + length];
+    String::from_utf8(bytes.to_vec()).unwrap_or_default()
+}
+
+#[derive(Debug)]
+struct SignatureRequestedEvent {
+    requester: Address,
+    payload_hash: [u8; 32],
+    path: String,
+    key_version: u32,
+    deposit: U256,
+    chain_id: U256,
+    algo: String,
+    dest: String,
+    params: String,
+}
+
+impl SignatureRequestedEvent {
+    fn encode_abi(&self) -> Vec<u8> {
+        let signature_requested_event_encoding = SignatureRequestedEncoding {
+            sender: self.requester,
+            payload: self.payload_hash.into(),
+            path: self.path.clone(),
+            keyVersion: self.key_version,
+            chainId: self.chain_id,
+            algo: self.algo.clone(),
+            dest: self.dest.clone(),
+            params: self.params.clone(),
+        };
+        signature_requested_event_encoding.encode_data()
+    }
+
+    pub fn generate_request_id(&self) -> [u8; 32] {
+        let abi_encoded = self.encode_abi();
+        alloy::primitives::keccak256(abi_encoded).into()
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::is_contract_call;
+    use alloy::primitives::Bytes;
+
+    #[test]
+    fn is_contract_call_detects_calldata() {
+        assert!(!is_contract_call(&Bytes::new()));
+        assert!(!is_contract_call(&Bytes::from(vec![0u8; 2])));
+        assert!(is_contract_call(&Bytes::from(vec![
+            0xa9, 0x05, 0x9c, 0xbb, 0x00
+        ])));
+    }
+
+    // TODO: Add more tests for event parsing
+}
